+++ src/dialogs/masterdialog.h
@@ -7,8 +7,12 @@ class MasterDialog : public QDialog
 {
 public:
     explicit MasterDialog(QWidget *parent = 0);
+
 public Q_SLOTS:
     virtual int exec();
+
+protected:
+    void resizeEvent(QResizeEvent *event);
 };
 
 #endif // MASTERDIALOG_H